Merge branch 'nr-polar-encoder-optimizations' of https://gitlab.eurecom.fr/oai/openairinterface5g into nr-polar-encoder-optimizations

Conflicts:
	openair1/PHY/NR_TRANSPORT/nr_dci_tools.c

+++ nfapi/open-nFAPI/nfapi/public_inc/nfapi_nr_interface.h
@@ -452,7 +452,7 @@ typedef struct {
   uint8_t rnti_type;
   uint8_t dci_format;
   /// Number of CRB in BWP that this DCI configures 
-  uint8_t n_RB_BWP;
+  uint16_t n_RB_BWP;
   uint8_t config_type;
   uint8_t search_space_type;  
   uint8_t aggregation_level;
